implementation of the linear formal network
this network is specifically designed to take in a consecutive iterator of elements (i.e. all consecutive pair of elements a, b follows the rule a.end == b.start)
and to return a consecutive iterator.

NB : two version have been made, one which allow to skip elements, the other that doesn't. When elements can be skipped, the starts/ends of the outputs won't correspond to the starts/ends of the inputs.
This behaviour (skip) is useful to implement WhiteSpace for example, but I tend to observe that the non-skip implementation is slightly faster (as the indexes doesn't have to be changed)

Future implentations ?

Maybe the index change on the implementation which allow skip can be optimized to equal the speed of the other one, then we could merge the two implementations into a single one)

We also could try to think about other behaviours that are compliant with the desired output type (i.e. consecutive elements iterator) and then implement them and add more configuration options